+++ services/api/src/controllers/subscribe.controller.ts
@@ -1,5 +1,4 @@
-import { RateLimit } from '../guards/limiter.guard';
-import { Controller, Post, UseGuards } from '../lib/router/router.decorators';
+import { Controller, Post } from '../lib/router/router.decorators';
 import { log } from '../lib/log';
 
 const logger = log.controller.from('subscribe');
@@ -10,7 +9,6 @@ const logger = log.controller.from('subscribe');
 @Controller('/subscribe')
 export class SubscribeController {
   @Post('')
-  @UseGuards(RateLimit('write'))
   async subscribe(req: Request) {
     let body: {
       email: string;